Created message system
Created message system, which allows users to send messeges to the
sellers and ask them about products or orders made by users.

Introduced chats and messages + theirs dtos. Chat represents 1
conversation and contains many messages, at least 1.

To write a message - create a chat - user has to click an appropriate
button in either Order or Product show-page. They select the title,
which is an enum, and give the message content.

Currently to check for new messages from other people, user has to
reload the message page. Right now, there are not any symbols on the
navbar telling if there is a unread message.

Messages within a chat are 'lazy-loaded' - there are 2 pages of messages
loaded (if exists) and the most recent page may not be full, because
there is ascending sort. When the user scrolls up the messages to the
top, new ones are loaded.

Chats are ordered by their last message date. Its also possible to
close/block the chat. If the chat is closed, users cannot write there
new messages, and only the user who closed the chat is able to unlock
it.

** Currently max message size is 255, as the default varchar in db and
that is not being validated during saving the messages to the db, so the
500 error can occur, it will be fixed later **