Rearrange the representation of products & orders
Right now, there are Products which cannot be removed and
AvailableProducts which can be removed. It allows people to view the
product page of a product which was deleted from store.

AvailableProducts are linekd to the actual product + they contain the
quantity of the available products, quantity of the bought products, and
boolean if the product was promoted. Those mechanisms will be introduced
soon.

Orders now represents bought product from 1 seller, if the user buys
products from more than 1 seller, he makes more than 1 order.