Simple products quantity logic
Now after making an order, the products' quantity changes according to
the bought products' quantity. Also on the product page and product list
there are displayed information about quantity of products sold.

If the product's quantity changes to 0, it expires - it's
AvailableProduct instance is deleted, but it's ExpiredProduct instance
is created. What it does is that the product is not displayed on the
product list and its being deleted from all the cart which were
contating it.
But it remains in the datebase as a ExpiredProduct and in future,
sellers will be able to set new quantity of the product, and recreate it
with the same properties.