Categories side column + extending category system
Added categories side column on the products page, so that user can
easily change the category.

Categories were changed, now they contain each other, what improves the
categorizing products.

Searching works through th whole category - itself and all the
descendants. To make it work, apart from the id, orderId was
provied, (probably will be renamed, because it can be confused with
Order class) and its used as a primary id, the old id is now just a
neccesary static id for db. OrderId can change, it happens when new
categories are added or (in the future) when the category order is
changed.

The order system is simple - children of a category must have an id
right after the parent. I hope it is a good way of searching through the
categories as we can just look for products with the category.orderId
within given range <main_category.id, main_category.last_child.id>

-- Work in progress --
Currently I'm working on category attributes - they will allow to set
some features of products such as size, color etc. and filter by them.
Unfortunatelly I commit this incomplete as I have to continue work on
this project on my laptop