Completed product attributes with filtering
I completed product attributes system, now it is possible to assign
attributes to a product and filter products by given attributes.
Now TODO is make some small changes to make it easier to add attributes
to categories, and change a little bit frontend to make it more clear
for users.

There are 2 types of attributes - numeric (NUMBER) and enumerated
(ENUM). The number ones can be filtered by MIN and MAX values, enum ones
are filtered by an enum value.

Now the way to finding products is simplified to only 1 method, it's
working using Criteria API and only neccesary Specifications are used.