Products sorting + small changes
Added functionality to sort products, right now available options are:
sorting by price, lowest first or highest first and a default one, if
neither chosen.

Right now if user moves to different category, the uncommon attributes
between the old category and a new one are erased. Because of that,
attributes from other categories are not affecting the results.