Fixed order making proccess
Fixed order making proccess so that now there are made all necessary
validations which check if the desired quantity of products in cart are
available.

Also created endpoint for users' profiles. Right now it redirects
requests to sellers page, where all the users products can be seen, but
I'm planning to make it somehow different